Split dipole_grid to make clear what each module does specifically
